+++ server/src/main/java/com/example/server/repositories/HeadphoneInfoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.example.server.entity.HeadphoneInfo;
 @Repository
-public interface HeadphoneInfoRepository extends JpaRepository<HeadphoneInfo, Integer> {
+public interface HeadphoneInfoRepository extends JpaRepository<HeadphoneInfo, Long> {
     
 }
